frontend-ts/frontendserver/requirements.txt
SQLAlchemy==2.0.40
sqlparse==0.4.4
Flask==3.0.2
flask-cors==6.0.1
Flask-Session==0.8.0
pandas==2.2.3
numpy==2.3.2
google==3.0.0
google-api-core==2.25.0rc0
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
requests==2.32.4
scipy==1.15.2

gunicorn
python-dotenv
gevent
functions_framework
psycopg2-binary

azure-identity
azure-storage-blob
azure-mgmt-compute
azure-mgmt-containerservice

PyYAML
kubernetes
pg8000